Rework debug UI structure and add base dialog for new reports
