+++ feature/smart-debugging/src/main/java/com/buzbuz/smartautoclicker/feature/smart/debugging/ui/dialog/live/conditiontry/TryImageConditionOverlayMenu.kt
@@ -14,46 +14,58 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.buzbuz.smartautoclicker.feature.smart.debugging.ui.live.eventtry
+package com.buzbuz.smartautoclicker.feature.smart.debugging.ui.dialog.live.conditiontry
 
 import android.util.Size
 import android.view.KeyEvent
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 
 import com.buzbuz.smartautoclicker.core.base.isStopScenarioKey
-import com.buzbuz.smartautoclicker.core.domain.model.scenario.Scenario
 import com.buzbuz.smartautoclicker.core.common.overlays.base.viewModels
 import com.buzbuz.smartautoclicker.core.common.overlays.menu.OverlayMenu
-import com.buzbuz.smartautoclicker.core.domain.model.event.ImageEvent
+import com.buzbuz.smartautoclicker.core.domain.model.condition.ImageCondition
+import com.buzbuz.smartautoclicker.core.domain.model.scenario.Scenario
 import com.buzbuz.smartautoclicker.feature.smart.debugging.R
-import com.buzbuz.smartautoclicker.feature.smart.debugging.databinding.OverlayTryEventMenuBinding
+import com.buzbuz.smartautoclicker.feature.smart.debugging.databinding.OverlayTryImageConditionMenuBinding
 import com.buzbuz.smartautoclicker.feature.smart.debugging.di.DebuggingViewModelsEntryPoint
-import com.buzbuz.smartautoclicker.feature.smart.debugging.ui.live.DebugOverlayView
-import com.buzbuz.smartautoclicker.feature.smart.debugging.uistate.ImageEventResultUiState
+import com.buzbuz.smartautoclicker.feature.smart.debugging.ui.dialog.live.uistate.ImageConditionResultUiState
+import com.buzbuz.smartautoclicker.feature.smart.debugging.ui.view.DebugOverlayView
+import com.google.android.material.slider.Slider
 
 import kotlinx.coroutines.launch
 
-class TryEventOverlayMenu(
+
+class TryImageConditionOverlayMenu(
     private val scenario: Scenario,
-    private val triedElement: ImageEvent,
+    private val imageCondition: ImageCondition,
+    private val onNewThresholdSelected: (Int) -> Unit,
 ) : OverlayMenu() {
 
     /** The view model for this dialog. */
-    private val viewModel: TryElementViewModel by viewModels(
+    private val viewModel: TryImageConditionViewModel by viewModels(
         entryPoint = DebuggingViewModelsEntryPoint::class.java,
-        creator = { tryElementViewModel() },
+        creator = { tryImageConditionViewModel() },
     )
 
-    private lateinit var viewBinding: OverlayTryEventMenuBinding
+    private lateinit var viewBinding: OverlayTryImageConditionMenuBinding
 
     override fun onCreateMenu(layoutInflater: LayoutInflater): ViewGroup {
-        viewBinding = OverlayTryEventMenuBinding.inflate(LayoutInflater.from(context))
+        viewBinding = OverlayTryImageConditionMenuBinding.inflate(LayoutInflater.from(context)).apply {
+            sliderThreshold.apply {
+                valueFrom = MIN_THRESHOLD
+                valueTo = MAX_THRESHOLD
+                value = imageCondition.threshold.toFloat()
+
+                addOnChangeListener(Slider.OnChangeListener { _, sliderValue, fromUser ->
+                    if (fromUser) viewModel.setThreshold(sliderValue.toInt())
+                })
+            }
+        }
 
         return viewBinding.root
     }
@@ -64,14 +76,16 @@ class TryEventOverlayMenu(
         lifecycleScope.launch {
             repeatOnLifecycle(Lifecycle.State.STARTED) {
                 launch { viewModel.displayResults.collect(::updateDetectionResults) }
+                launch { viewModel.thresholdText.collect(viewBinding.valueThreshold::setText) }
             }
         }
 
-        viewModel.startTry(context, scenario, triedElement)
+        viewModel.startTry(context, scenario, imageCondition)
     }
 
     override fun onStop() {
         viewModel.stopTry()
+        onNewThresholdSelected(viewModel.getSelectedThreshold())
     }
 
     override fun getWindowMaximumSize(backgroundView: ViewGroup): Size {
@@ -102,8 +116,8 @@ class TryEventOverlayMenu(
         return true
     }
 
-    private fun updateDetectionResults(results: ImageEventResultUiState?) {
-        (screenOverlayView as? DebugOverlayView)?.setResults(results?.detectionResults ?: emptyList())
-        viewBinding.textResult.text = results?.eventText
+    private fun updateDetectionResults(results: ImageConditionResultUiState?) {
+        (screenOverlayView as? DebugOverlayView)?.setResults(results?.let { listOf(it) } ?: emptyList())
+        viewBinding.valueResult.text = results?.resultText
     }
 }
