--- feature/smart-debugging/src/main/java/com/buzbuz/smartautoclicker/feature/smart/debugging/ui/report/DebugReportDialog.kt
@@ -1,50 +1,0 @@
-/*
- * Copyright (C) 2023 Kevin Buzeau
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.buzbuz.smartautoclicker.feature.smart.debugging.ui.report
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-
-import com.buzbuz.smartautoclicker.core.common.overlays.base.viewModels
-import com.buzbuz.smartautoclicker.core.common.overlays.dialog.OverlayDialog
-import com.buzbuz.smartautoclicker.feature.smart.debugging.R
-import com.buzbuz.smartautoclicker.feature.smart.debugging.databinding.DialogDebugReportBinding
-import com.buzbuz.smartautoclicker.feature.smart.debugging.di.DebuggingViewModelsEntryPoint
-
-import com.google.android.material.bottomsheet.BottomSheetDialog
-
-/** Displays the content of the current debug report. */
-class DebugReportDialog : OverlayDialog(R.style.AppTheme) {
-
-    /** View model for this dialog. */
-    private val viewModel: DebugReportModel by viewModels(
-        entryPoint = DebuggingViewModelsEntryPoint::class.java,
-        creator = { debugReportModel() },
-    )
-
-    /** ViewBinding containing the views for this dialog. */
-    private lateinit var viewBinding: DialogDebugReportBinding
-
-    override fun onCreateView(): ViewGroup {
-        viewBinding = DialogDebugReportBinding.inflate(LayoutInflater.from(context))
-        return viewBinding.root
-    }
-
-    override fun onDialogCreated(dialog: BottomSheetDialog) {
-
-    }
-}
