+++ feature/smart-debugging/src/main/res/menu/menu_debug_report.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+* Copyright (C) 2025 Kevin Buzeau
+*
+* This program is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, either version 3 of the License, or
+* (at your option) any later version.
+*
+* This program is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* along with this program.  If not, see <http://www.gnu.org/licenses/>.
+-->
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/page_overview"
+        android:enabled="true"
+        android:icon="@drawable/ic_debug_overview"
+        android:title="@string/menu_item_debug_report_overview"/>
+    <item
+        android:id="@+id/page_timeline"
+        android:enabled="true"
+        android:icon="@drawable/ic_debug_timeline"
+        android:title="@string/menu_item_debug_report_timeline"/>
+</menu>
